Added JWT tokens. Did the token generation, sert the ID as the subject and added fields like name, email, and phone number. Added the authentication filter which checks the header and token validity and returns the extracted token. Added helper function so all controllers can get userID from token and compare them before doing sensitive API calls. Fixed small error when editing emails for user profiles.

+++ Backend/seasonal_job_matching/src/main/java/grad_project/seasonal_job_matching/controller/ResumeController.java
@@ -3,6 +3,7 @@ package grad_project.seasonal_job_matching.controller;
 import java.util.Map;
 import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -15,9 +16,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import grad_project.seasonal_job_matching.services.ResumeService;
 import grad_project.seasonal_job_matching.dto.responses.ResumeResponseDTO;
+import grad_project.seasonal_job_matching.security.CurrentUserService;
 import grad_project.seasonal_job_matching.dto.requests.ResumeCreateDTO;
 import grad_project.seasonal_job_matching.dto.requests.ResumeEditDTO;
-
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 
 @CrossOrigin(origins = "*")
@@ -26,11 +28,12 @@ import jakarta.validation.Valid;
 public class ResumeController {
 
     final private ResumeService resume_service;
-    //final private UserService user_service; will add once sessions are created
+    final private CurrentUserService currentUserService;
 
-    public ResumeController(ResumeService resume_service){
+    public ResumeController(ResumeService resume_service, CurrentUserService currentUserService){
         this.resume_service = resume_service;
-        //this.user_service = user_service;
+        this.currentUserService = currentUserService;
+        
     }
 
     // @GetMapping
@@ -39,7 +42,11 @@ public class ResumeController {
     // }
 
     @GetMapping("/{id}")//using user ID, if has no resume returns 500, could add try catch to avoid this
-    public ResponseEntity<?> findByID(@PathVariable long id){
+    public ResponseEntity<?> findByID(@PathVariable long id, HttpServletRequest request){
+        Long currentUserId = currentUserService.getCurrentUserId(request);
+        if (currentUserId == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        if (currentUserId != id) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
         Optional<ResumeResponseDTO> resume = resume_service.findResumeByUserId(id);
         if(resume.isEmpty()){
             return ResponseEntity.ok("Resume not found!");
@@ -50,7 +57,10 @@ public class ResumeController {
     }
 
     @PostMapping("/{userId}")
-    public ResponseEntity<?> createResume(@Valid @RequestBody ResumeCreateDTO dto, @PathVariable long userId){//if user is from mobile than type is jobseeker, else it is employer  
+    public ResponseEntity<?> createResume(@Valid @RequestBody ResumeCreateDTO dto, @PathVariable long userId, HttpServletRequest request){//if user is from mobile than type is jobseeker, else it is employer  
+        Long currentUserId = currentUserService.getCurrentUserId(request);
+        if (currentUserId == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        if (currentUserId != userId) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         try {
             ResumeResponseDTO resume = resume_service.createResume(dto,userId);
             return ResponseEntity.ok()
@@ -66,7 +76,11 @@ public class ResumeController {
 
 
     @PatchMapping("/{userId}")
-    public ResponseEntity<?> editResume(@PathVariable long userId,@Valid @RequestBody ResumeEditDTO dto){
+    public ResponseEntity<?> editResume(@PathVariable long userId,@Valid @RequestBody ResumeEditDTO dto, HttpServletRequest request){
+        Long currentUserId = currentUserService.getCurrentUserId(request);
+        if (currentUserId == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        if (currentUserId != userId) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
         try {
             ResumeResponseDTO resume = resume_service.editResume(dto, userId);
             return ResponseEntity.ok()
@@ -82,13 +96,11 @@ public class ResumeController {
 
 
     @DeleteMapping("/{userId}")
-    public ResponseEntity<String> deleteResume(@PathVariable Long userId){
-        // if ("Resume found!".equals(findByID(userId).getBody())) {
-        //     resume_service.deleteResume(userId);
-        //     return ResponseEntity.ok("Resume deleted successfully!");
-        // }else{
-        //     return ResponseEntity.ok("Resume not found!");
-        // }
+    public ResponseEntity<String> deleteResume(@PathVariable Long userId, HttpServletRequest request){
+        Long currentUserId = currentUserService.getCurrentUserId(request);
+        if (currentUserId == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        if (currentUserId != userId) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
         if (resume_service.findResumeByUserId(userId).isPresent()) { 
         resume_service.deleteResume(userId);
         return ResponseEntity.ok("Resume deleted successfully!");
